Cleaned spriter++ warnings and TODO

+++ ext/spriter/variable/variableinstancenameandidmap.cpp
@@ -7,53 +7,53 @@
 namespace SpriterEngine
 {
 
-	VariableInstanceNameAndIdMap::VariableInstanceNameAndIdMap()
-	{
-	}
+    VariableInstanceNameAndIdMap::VariableInstanceNameAndIdMap()
+    {
+    }
 
-	VariableInstanceNameAndIdMap::~VariableInstanceNameAndIdMap()
-	{
-		for (auto& it : variables)
-		{
-			delete it;
-		}
-	}
+    VariableInstanceNameAndIdMap::~VariableInstanceNameAndIdMap()
+    {
+        for (auto& it : variables)
+        {
+            delete it;
+        }
+    }
 
-	UniversalObjectInterface *VariableInstanceNameAndIdMap::getVariable(int variableIndex)
-	{
-		if(variableIndex<variables.size())
-		{
-			return variables.at(variableIndex);
-		}
-		else
-		{
-			// error
-			return 0;
-		}
-	}
+    UniversalObjectInterface *VariableInstanceNameAndIdMap::getVariable(int variableIndex)
+    {
+        if(unsigned(variableIndex) < variables.size())
+        {
+            return variables.at(variableIndex);
+        }
+        else
+        {
+            // error
+            return 0;
+        }
+    }
 
-	UniversalObjectInterface *VariableInstanceNameAndIdMap::getVariable(std::string variableName)
-	{
-		auto it = variableNameMap.find(variableName);
-		if (it != variableNameMap.end())
-		{
-			return (*it).second;
-		}
-		else
-		{
-			return 0;
-			// error
-		}
-	}
+    UniversalObjectInterface *VariableInstanceNameAndIdMap::getVariable(std::string variableName)
+    {
+        auto it = variableNameMap.find(variableName);
+        if (it != variableNameMap.end())
+        {
+            return (*it).second;
+        }
+        else
+        {
+            return 0;
+            // error
+        }
+    }
 
-	void VariableInstanceNameAndIdMap::pushBackVariable(Variable *variable)
-	{
-		variables.push_back(variable->getNewObjectInfoInstance(true));
-		variableNameMap[variable->getName()] = variables.back();
-	}
+    void VariableInstanceNameAndIdMap::pushBackVariable(Variable *variable)
+    {
+        variables.push_back(variable->getNewObjectInfoInstance(true));
+        variableNameMap[variable->getName()] = variables.back();
+    }
 
-	bool VariableInstanceNameAndIdMap::isEmpty()
-	{
-		return variables.empty();
-	}
+    bool VariableInstanceNameAndIdMap::isEmpty()
+    {
+        return variables.empty();
+    }
 }
